+++ components/common/Form/Form.module.scss
@@ -1,4 +1,4 @@
-.form {
+.Form {
     margin: 1rem;
     padding: 1rem;
     display: flex;
