+++ components/common/Form/FormItem/FormItem.tsx
@@ -3,6 +3,6 @@ import styles from "./FormItem.module.scss";
 
 export const FormItem:FC = ({children}) => {
   return (
-    <div className={styles.FromItem}>{children}</div>
+    <div className={styles.FormItem}>{children}</div>
   )
 }
