+++ components/common/Form/FormItem/FormSelectField.tsx
@@ -1,6 +1,5 @@
 import React, { FC } from "react";
 import { FieldValues, UseFormRegister } from "react-hook-form";
-import styles from "../../styles/Form.module.scss";
 import { ISelect } from "../../../../types/ISelect";
 import { FormItem } from "./FormItem";
 
@@ -15,7 +14,7 @@ export const FormSelectField: FC<IFormSelectField> = ({ register, item }) => {
       <label>{item.label}</label>
       <select
         {...register(!!item.name ? item.name : item.label)}
-        className={styles.form__item}
+        
       >
         {item.choices.map((c, j) => (
           <option key={j} value={c.key}>
