+++ components/common/Read/Read.module.scss
@@ -1,4 +1,4 @@
-.read {
+.Read {
     margin: 1rem;
     margin-top: 5rem;
     padding: 3rem;
@@ -10,6 +10,8 @@
     align-items: center;
     justify-content: center;
     &__icons {
+        display: flex;
+        justify-content: space-between;
         &__edit {
             margin: 0.3rem;
             padding: 0.5rem;
