+++ components/common/Read/Read.tsx
@@ -1,5 +1,4 @@
-
-import {IExpense} from "../../../types/IExpense"
+import { IExpense } from "../../../types/IExpense";
 
 import React, { FC } from "react";
 import styles from "./Read.module.scss";
@@ -9,23 +8,25 @@ import { Category, Merchant } from "@prisma/client";
 
 export interface IRead {
   item: Category | Merchant | IExpense;
-  onDelete: () => void;
+  onDelete?: () => void;
   handleEdit: () => void;
 }
 
 export const Read: FC<IRead> = ({ item, children, handleEdit, onDelete }) => {
   return (
     <div className={styles.Read}>
-      <FontAwesomeIcon
-        onClick={() => handleEdit()}
-        className={styles.Read__edit}
-        icon={faPencil}
-      />
-            <FontAwesomeIcon
-        onClick={onDelete}
-        className={styles.read__icons__delete}
-        icon={faTrash}
-      />
+      <div className={styles.Read__icons}>
+        <FontAwesomeIcon
+          onClick={() => handleEdit()}
+          className={styles.Read__icons__edit}
+          icon={faPencil}
+        />
+        <FontAwesomeIcon
+          onClick={onDelete}
+          className={styles.Read__icons__delete}
+          icon={faTrash}
+        />
+      </div>
       {children}
     </div>
   );
