+++ components/merchants/ReadMerchant.tsx
@@ -1,9 +1,10 @@
 import { Category, Merchant } from "@prisma/client";
 import React, { FC } from "react";
-import styles from "../../styles/Read.module.scss";
+import styles from "./ReadMerchant.module.scss";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPencil } from "@fortawesome/free-solid-svg-icons";
 import { useRouter } from "next/router";
+import { Read } from "../common/Read/Read";
 
 interface IReadMerchant {
   merchant: Merchant | undefined;
@@ -19,17 +20,13 @@ export const ReadMerchant: FC<IReadMerchant> = ({ merchant }) => {
     router.push(`/merchants/${merchant.id}/edit`);
   };
   return (
-    <div className={styles.read}>
-      <FontAwesomeIcon
-        onClick={() => handleEdit()}
-        className={styles.read__edit}
-        icon={faPencil}
-      />
-      <h3 className={styles.read__name}>{merchant.name}</h3>
-      <h6 className={styles.read__id}>id: {merchant.id}</h6>
-      <p className={styles.read__last}>Spent last month: 0zł</p>
-      <p className={styles.read__current}>Spent this month: 0zł</p>
-    </div>
+    <Read item={merchant} handleEdit={handleEdit}>
+      <div className={styles.ReadMerchant}>
+        <h3 className={styles.ReadMerchant__name}>{merchant.name}</h3>
+        <h6 className={styles.ReadMerchant__id}>id: {merchant.id}</h6>
+        <p className={styles.ReadMerchant__last}>Spent last month: 0zł</p>
+        <p className={styles.ReadMerchant__current}>Spent this month: 0zł</p>
+      </div>
+    </Read>
   );
 };
-
